agent classes separated
