--- bankcraft/Motivation.py
@@ -1,9 +1,0 @@
-class Motivation:
-    def __init__(self):
-        self.mtv_dict = { 'hunger' : 1,
-                         'fatigue' : 1,
-                         'social' : 1, 
-                         'consumer_needs' : 1
-        }
-
-
