+++ bankcraft/agent/general_agent.py
@@ -1,0 +1,9 @@
+from mesa import Agent
+
+
+class GeneralAgent(Agent):
+    def __init__(self, unique_id, model):
+        super().__init__(unique_id, model)
+
+    def step(self):
+        pass
