+++ bankcraft/motivation.py
@@ -1,0 +1,7 @@
+class Motivation:
+    def __init__(self):
+        self.motivation_dict = {'hunger': 1,
+                                'fatigue': 1,
+                                'social': 1,
+                                'consumer_needs': 1
+                                }
